changed integerfields for models.py

+++ apps/login/models.py
@@ -68,7 +68,7 @@ class Users(models.Model):
     email = models.CharField(max_length=60, unique=True)
     password = models.CharField(max_length=50)
     username = models.CharField(max_length=60, unique=True)
-    userlevel = models.PositiveSmallIntegerField(default=0)
+    userlevel = models.IntegerField(default=0)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
